Move window provider and back provider
