+++ src/pages/MapPage/MapPage.js
@@ -19,7 +19,7 @@ import useStep from '../../utils/hooks/useStep'
 import TagDetailDialog from '../../components/TagDetail/TagDetailDialog'
 import FilterFab from '../../components/Filter/FilterFab'
 import LocationFab from '../../components/LocationFab'
-import WindowBackProvider from '../../components/WindowBackProvider'
+import WindowBackProvider from '../../utils/WindowBackProvider'
 
 export default function MapPage(props) {
   const { signOut, deny, guest } = props
